(BKR-496) Fix create_tmpdir_on() for all platforms
This fix directly addresses macOS and Windows, but provides a general solution by using host.user_get() to check if user (or host's default) exists on host. Also validates the group to accept a passed value or fall back to host default (instead of just using username for group). Spec tests and docs are expanded to cover addition of group argument.