(BKR-462) Check existence of SSH key files
Host#do_rsync_to unsafely assumed that the first entry in the options hash SSH keys was valid. This change makes it check through the array of configured keys until it finds one that actually exists in the filesystem.

Note: the spec tests somehow mess this up and the lambda returns an Array. It works just fine in acceptance and I cannot reproduce the issue in a standalone test case. Weird.