(BKR-462) Fixes rsync issues around rsync not installed on remote host (#1538)
* (BKR-462) rsync_to fails if rsync is not installed on the remote host

Also adds solaris to the list of hosts on which rsync is not installed by default.

* (BKR-462) Adds #uninstall_package host helper

The #install_package test uses rsync as an example package, but doesn't clean up after itself on hosts that don't have rsync installed by default. This lack of cleanup was contaminating later tests for rsync that expected rsync. This change adds an uninstall_package host helper to match the install_package helper, and calls this helper to clean up after install_package tests. It also is applied to other tests to replace command-line package removal, which didn't work for solaris hosts.

* (BKR-462) Check existence of SSH key files

Host#do_rsync_to unsafely assumed that the first entry in the options hash SSH keys was valid. This change makes it check through the array of configured keys until it finds one that actually exists in the filesystem.

Note: the spec tests somehow mess this up and the lambda returns an Array. It works just fine in acceptance and I cannot reproduce the issue in a standalone test case. Weird.

* (BKR-462) CHANGELOG.md

* (BKR-462) Use safe logic to identify existing SSH keys

Also fixup spec tests so they don't include the same code copy-paste as the method under tests.

* (BKR-462) Fix remaining spec tests

D'oh.

* (BKR-462) *ahem*

* (BKR-462) Remove redundant acceptance test for #install_package host helper

This acceptance test is redundant, since #install_package is just an alias for Host#install_package. The acceptance test as written was causing downstream test issues by installing rsync and not cleaning up after itself. Spec tests are added to cover this behavior.

* (BKR-462) Clean up rsync_to_test.rb

Previously, rsync_to_test.rb assumed certain hosts did or did not have rsync installed by default. However, due to sloppy tests for Host Helpers#install_package (now removed), rsync was being installed on some of those hosts which were believed not to have it. Now, functionality is exercised on all hosts by ensuring rsync is uninstalled or installed as appropriate, and properly cleaned up afterward.

* (BKR-462) Workaround for uninstall_package on Solaris 10

See BKR-1502 to fix.

* (BKR-462) Typo

* (BKR-462) Packages are not supported on OSX

* (BKR-462) Typo :(

* (BKR-462) And don't check_for_package anymore

* (BKR-462) Fix rsync_to_test on OSX and Solaris 10

* (BKR-462) Fix rsync dependency in create_remote_file_test on Windows, OSX, and Solaris

Includes a workaround for BKR-1502

* (BKR-462) Mark Windows rsync issues

* (BKR-462) Missed one during rebase