(BKR-1541) Preserve host hash from beaker exec run
Prior to this change, we were losing state that had been set up when
running our pre-suite. Unfortunately, this oversight wasn't felt because
beaker was loading these defaults automatically for every beaker
invocation. In beaker 4, that functionality was removed, and we found
that state between beaker subcommand invocations was unfortunately not
preserved. This commit adds in an optional flag `--preserve-state` that
users can pass into a `beaker exec` invocation. This will update the
beaker options file to amend the host has to include any new or updated
information. This is particularly relevant for paths like
privatebindir` that are referenced in tests.