(BKR-1549) Recursively remove unpersisted subcommand options
Certain option keys (like :logger, for a logger object), should never be
persisted in a subcommand options yaml file. SubcommandUtil was already
removing these keys from the top level of any options hash before
persisting it, but it wasn't recursing into child values.

If a child value (like the HOSTS hash, written during provisioning)
contained a :logger entry, that entry was persisted to the options file
as a string. This string representation of the logger would subsequently
be used while trying to ssh to a host, raising an exception.

This change updates sanitize_options_for_save so that it recursively
removes unwanted keys from any configs that are about to be persisted.
It also adds :password_prompt to the list of ignored options, to avoid
persisting Net::SSH password prompt objects.