(BKR-1559) Do not mirror profile.d on Debian
This mirroring is not needed since directly executed SSH commands ignore
profile.d. This means it's sufficient to set ~/.ssh/environment.

The format is another issue with this function. In ~/.ssh/environment
it'll be PATH=PATH:/opt/puppetlabs/bin but this breaks the PATH in an
actual shell.

This only removes it on Debian because I've verified that works and I
don't have any SLES-based machines nor knowledge.