(BK-1562) return root considerations for appending on nexus devices
BK-1556 changes refactored the appending of strings and it was missed when the user is a root user, which caused an EOF error on some use cases, i.e. when running guestshell as a root user.

This commit adds in the root user check to only append `"` for non-root users.