(MOB) fix fibonacci repetition sleep issue
Previously, if the block passed to #repeat_fibonacci_style_for
returned true, meaning it was done repeating, the method still
slept as if it had not succeeded, wasting time. This fix makes
execution skip the sleep if the block has reported success,
returning execution to the user a little quicker.