(BKR-1604) Add FIPS detection method to Host
This commit adds a method that reports if FIPS mode is enabled for a
host. Currently, we only test on Redhat 7 with FIPS mode, which
indicates its state via a file. If we eventually expand support to
testing with FIPS on more operating systems, those cases and their
detection methods should be added. For now, all other platforms simply
return false.