(RE-12690) Raise scp error on failure
Filesync tests are experiencing a recurring transient issue where scp fails and
we end up with a corrupted .list file on ubuntu hosts. This should give us more
info as to *why* the scp is failing by including the actual error output. This
should also prevent processes from continuing after the error, which previously
was resulting in further-downstream errors.