(QENG-7422) Add wait_time and max_connection_tries to exec.reboot
Prior to this commit there was no way to customize the behavior of the
Host::Unix::Exec.reboot method. This meant that if you had a host that
took a particularly long time to come back from a reboot, your tests
would fail.

This commit adds two parameters, `wait_time` and `max_connection_tries`,
which will modify how long to wait before checking on the host after
rebooting it and how many times to attempt connecting to the host after
the reboot. As noted in the method comment, the `max_connection_tries`
uses a fibbonacci backoff so increasing it will not increase the time
spent waiting for reconnection in a linear way.

Additionally, I have added specs for these parameters and mocked out
`Kernel.sleep` in the specs so that they run faster.