(QENG-7422) Add retries to uptime check
Prior to this commit if the `uptime` command did not reset after the
provided `wait_time`, we would assume that the host wasn't going to
reboot.
This commit adds an `uptime_retries` parameter to the
`Unix::Exec.reboot` method that specifies how many times Beaker should
attempt to check the result of the `uptime` command before giving up on
the host.