Run host generation tests in Travis
This modifies the `:travis` rake task to run all available tests in the
Rakefile, including host generation tests and Beaker subcommand tests.
This ensures that any changes that may break Beaker host generation or
subcommands will be caught before the change is merged.

This additional bumps the required Ruby version to 2.4, since Beaker
uses the safe navigation operator `&.` which was introduced in Ruby 2.3,
and in order to use Ruby 2.3 the required pry and pry-byebug versions
that are compatible require Ruby 2.4.

It also updates the Travis ruby matrix to include the lowest supported
version - Ruby 2.4.