Handle systems going back in time after reboot
Updates the UNIX reboot method to handle systems that go back in time
after reboot.

This can happen often on virtual machines that are running on a heavily
loaded system.

The error handling had to be loosened to allow the loop to handle
whatever bizarre circumstances get thrown at it by the underlying
system.

Fixes #1674