Use chrony instead of ntpdate on Fedora
The ntpdate package is no longer available starting with Fedora 34, the
ntpdate binary is now provided by the ntpsec package.

Since chrony seems to be the preferred way of synchronizing time, use it
on all Fedora. We tested this internally and the chrony package comes
preinstalled since at least Fedora 30.