(maint) Use forward slashes for Windows-OpenSSH
When using cygwin bash.exe, backslashes are handled differently depending on
whether we're using cygwin or Windows-OpenSSH's sshd.

    $ ssh Administrator@win32-openssh 'cmd /c echo c:\\Windows'
    c:Windows

    $ ssh Administrator@cygwin 'cmd /c echo c:\\Windows'
    c:\Windows

Note that Win32-OpenSSH lost the backslashes. This issue prevents beaker from
installing puppet-agent on Windows 11 using Win32-OpenSSH, as it scp's a bat
file to the Windows host:

    $ scp /tmp/install-puppet-msi-2022-06-22_18.18.04.bat20220622-2096997-1w3xtbd host.example.com:C:\Windows\TEMP\install-puppet-msi-2022-06-22_18.18.04.bat

But the resulting path doesn't exist:

    $ cmd.exe /c "C:\Windows\TEMP\install-puppet-msi-2022-06-22_18.18.04.bat"
    'C:\Windows\TEMP\install-puppet-msi-2022-06-22_18.18.04.bat' is not recognized as an internal or external command, operable program or batch file.

To avoid these problems, convert backslashes to forward slashes when using
Windows-OpenSSH. This way the scp and execute steps refer to the same file. This
is the same solution used when using bitvise and `Beaker::Host:PSWindows` since
it's `scp_separator` is '/'[1]

[1] https://github.com/voxpupuli/beaker/blob/76a974f448bc49cdab69792d95ae606d2d868f50/lib/beaker/host/pswindows.rb#L28