(maint) Fix ArgumentError
If `determine_ssh_server` returned an unknown type, then beaker failed to raise
an ArgumentError. When using `raise` you can pass two arguments, the class to
raise and the message, or pass an instance of an exception, e.g.
ArgumentError.new("message").

This also adds tests for the other cases.