Create ~/.ssh on Windows if it doesn't exist
Beaker executed `mkdir -p "~/.ssh"` on Windows hosts as part of the prebuilt
steps sequence. However, when using cygwin bash, the double quotes behave like
single quotes, disabling tilde expansion:

   $ realpath "~/.ssh"
   /home/Administrator/~/.ssh

So '~' was a directory containing the '.ssh' directory. This was never noticed
because all of our VM templates already have the expected ~/.ssh directory. But
if the directory doesn't exist, then later steps would fail trying to write to
~/.ssh/environment.

This commit changes beaker behavior in the case that the path starts with tilde
and does not contain a space, so it's safe to not quote the path.

See commit 54b35b7603d25f4b38d11d0173ecd489ee96cad4 which double quoted unix &
windows, which caused a regression that was fixed in 9288623e6e95e5b144d565b2206aad994d15bdcb.