Add to list of permitted classes and allow aliases
When trying to load hosts_preserved.yml, all of these classes need to be able to be loaded. We should probably serialize fewer things to this file, but in the meantime, this allows things to work. This also uses strings for the list, since not all the classes are loaded and available at the time we are declaring the array. This also allows aliases, since we are using those in the hosts_preserved.yml file.

This also adds a test to the spec to verify we can load a representative hosts_preserved.yml.