Remove of =~ which Performance/RedundantMatch didn't find
In Ruby 3.2 Kernel.=~ has been removed so it should be removed as much
as possible. This replaces it with match? or start_with? when
appropriate. The Performance/RedundantMatch cop can find quite a few of
them, but not all.

Note there are still uses remaining, for example when the capture groups
were used.