(maint) Removes open_uri_redirections
Prior to Ruby 2.4, the OpenURI module did not natively support
HTTP => HTTPS redirections, necessitating a third-party gem,
open_uri_redirections.

However, redirection functionality was added in Ruby 2.4:

https://github.com/ruby/ruby/commit/393ecc9f10016828a3713438ae216dc39b965364

And the `Kernel#open` method was deprecated in Ruby 2.7 and removed
entirely in Ruby 3.0 in favor of the `URI#open` method:

https://bugs.ruby-lang.org/issues/15893

This commit removes the open_uri_redirections gem from the WebHelpers
module and gemspec and replaces `Kernel#open` with `URI#open`.