Correct Amazon package manager
73edd1a backported Amazon Linux support from master to 4.x. However,
because the commit did not apply cleanly, the install_package method
mistakenly uses Yum on Amazon Linux 2023.

DNF is the default package manager for Amazon Linux 2023 and Beaker 5.x
correctly uses DNF in the install_package method.

This commit changes Amazon Linux 2023 to use DNF instead of Yum in the
install_package method.