Default to DNF for Fedora
Prior to this commit, the install and uninstall methods for Fedora
defaulted to using DNF as its package manager for versions greater than
or equal to 22 but less than 40, and Yum for every other version.

This commit configures Beaker to use DNF for every version of Fedora, as
DNF has been the default package manager for Fedora for the last 8
years.