Ensure platform is Beaker::Platform in tests
The regular option parser guarantees this and it sort of works because
Beaker::Platform can behave as a string (since it inherits from it).
When you do want to use extended features, it fails.

This enhances the helper to guarantee it's the correct class instance.
Then there are many changes to the tests to ensure a valid platform is
passed in.