Simplify the wildcard *.pkg search
Note that a `for-loop` was suggested to do the wildcard search.  This
change removes the `for-loop` in favour of a `find` for several reasons:

(1) the original for-loop, unexpectedly, requires the if-statement to
bomb out if no valid '*.pkg' file is found.  This happens even though a
for-loop will normally return a non-zero exit code if nothing is found.

(2) the find implementation is a good bit simpler and fewer lines.

Changing from a for-loop to a find to shorten the code and hopefully simplify.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>